Add external argparse instance support
While subcommands have the support of an external `Argparse` instance,
main commands do not have this flexibility. This commit aims to
implement that behavior.

I tried to modify minimum amount of code parts. I did not use
`_CopiedArgumentParser` because it seemed to be only usable in
subcommand structure, since no other code part uses `_parser_class`.

I'm not sure if it is a good idea, but with the initial version, the
wrapper supports both an external instance and annotation argument
definitions, as such can be seen in the tests.

This commit fixes #5.